Preregistration
This is the preregistration of the NOR-SOLIDARITY trial analyses. It follows the statistical analysis plan version 1.0 dated 4 Jan 2021 and prespecifies all analyses to be performed in the main part of the trial. All code has been produced without knowing any results. The code will be run with the true allocation once the statistical analysis plan has been signed.